first attempt at creating a single mapping document
both the dtbook-to-epub3 and epub3-to-dtbook conversion paths will
be described by this single mapping document.

A fourth column containing the dtbook element names has been added
but the descriptions (i.e. what classes must have what values etc
for the template to match) has not been fixed and all the descriptions
in the table cells must be updated.