experimental script: EPUB3 ASCIIMath to MathML
This is a first attempt at tackling #54.

Script is deactivated by default in META-INF/catalog.xml, but works relatively well.
(for testing purposes you can open the JAR for this module and edit catalog.xml to enable this script)
We need to discuss in further detail how to include MathML, and this script can be
used as a basis for discussions.

asciimath spans have been added to the test documents. No tests have been added for
the experimental script, since it depends on asciimath-utils which must be run inside
of Pipeline 2. It might be possible to write tests for it if we get integrations tests
working. See: #86

Also:
 - changed URI component in XML catalog entries from "dtbook-to-epub3" to "epub3-dtbook-migrator"
 - added asciimath to allowed classes in epub3-to-dtbook
 - added XSLT tests in dtbook-to-epub3 and epub3-to-dtbook to make sure asciimath class is preserved
 - fixed links in test documents to make them valid