make nav toc based on single-html to allow for chapters wrapped in parts
there's no way to infer a structure like this based on the multi-html representation:
 * frontmatter-1.xhtml
 * chapter-2.xhtml
 * part-3.xhtml
   * chapter-4.xhtml
   * chapter-5.xhtml
 * chapter-6.xhtml
 * part-7.xhtml
   * chapter-8.xhtml
   * chapter-9.xhtml
 * backmatter-10.xhtml

this solution makes the nav toc based on the single-html representation, and then fixes
all the hrefs afterwards. this is possible because we're using the nordic guidelines which
dictates that all ids must be unique across the entire publication, and because the script
uses the same ids in the single-html representation as in the multi-html representation.

fixes #134