allow rear-/footnotes at depth=3
fixes #296

If for instance you place a `<section epub:type="rearnotes">` inside a
`<body epub:type="bodymatter chapter">`, and the chapter is inside
a part, then the rearnotes section will end up at depth=3
(i.e. with a `<h3>`), which used to be invalid. This commit allows
the types "rearnotes" and "footnotes" at depth=3.