fixes fail-on-error issue. adds tests for invalid books.
* set version to 1.1.0 (features have been added since 1.0.0, could
arguable increase major number instead)
* added `fail-on-error-status.xpl` implementing `px:nordic-fail-on-
error-status` as a helper step to determine validation status when fail-
on-error is false. If there is no file with either the `xml`, `xhtml` or
`epub` file extension in the output directory, the script will report
failure even though fail-on-error is false.
* added required option `fail-on-error` to all script steps (they were
optional with a default set to 'true'). The `fail-on-error` option were
not properly passed on to all sub-steps when running the scripts, which
caused it to be ignored in some cases. Should fix #321
* added tests to make sure that the optional `fail-on-error` option on
the conversion scripts are working
* added tests for validation scripts to make sure that they fail when
validating invalid books
* as part of the `fail-on-error` tests, added invalid versions of the
test DTBook, single HTML, unzipped EPUB and zipped EPUB
* removed `fail-on-error` option from nordic-dtbook-validator and
nordic-html-validator scripts; it doesn't make sense on validation
scripts
* added `dtbook-store.step.xpl` implementing `px:nordic-dtbook-
store.step` to properly and consistently store DTBooks